producers/requirements.txt
kafka-python==2.0.2
faker==20.1.0
numpy==1.24.3

